Move namespace. Remove metrics for the unit test since it will have user_events tests for logs and traces as well.
